Rename util/getHideCaretAnimationName -> recipes/hideCaret.
