+++ src/components/LayoutTree.tsx
@@ -18,6 +18,7 @@ import { HOME_PATH } from '../constants'
 import testFlags from '../e2e/testFlags'
 import useFauxCaretCssVars from '../hooks/useFauxCaretCssVars'
 import useSortedContext from '../hooks/useSortedContext'
+import hideCaret from '../recipes/hideCaret'
 import attributeEquals from '../selectors/attributeEquals'
 import calculateAutofocus from '../selectors/calculateAutofocus'
 import findDescendant from '../selectors/findDescendant'
@@ -36,7 +37,6 @@ import scrollTopStore from '../stores/scrollTop'
 import viewportStore from '../stores/viewport'
 import { appendToPathMemo } from '../util/appendToPath'
 import equalPath from '../util/equalPath'
-import hideCaret from '../util/getHideCaretAnimationName'
 import hashPath from '../util/hashPath'
 import head from '../util/head'
 import isRoot from '../util/isRoot'
