+++ src/recipes/hideCaret.ts
